requirements.txt
boxsdk
mysql-connector